move notifications to its own module
