+++ proj/src/main/webapp/appconfig.js
@@ -8,3 +8,4 @@ exports.firebaseConfig = {
     appId: "1:336825043126:web:1256d7b08f1c8daa93be17",
     measurementId: "G-VJQS6TEYGV"
 };
+exports.serverKey = 'key=AAAATmxWSLY:APA91bFngBwWTr1GyOT_bxAK1aiCtj4wPc8QT7NvrAi1mwkYYxSQGUn7ZeHKarx_Oc0OH_7qe61VaWFdKybizER9G77xP9Y3f77R9l4t2095CWSCPx77xM7nprK6jWihdjlAarRb2Zd3';
